+++ admin/src/components/Customization/Products/ProductForm/ProductItemsModal/ProductItemsList/ProductItemsList.tsx
@@ -2,14 +2,13 @@
 
 import { getAllAvailableItems } from '@/api';
 import ItemCard from '@/components/Customization/ItemCard/ItemCard';
-import { Button, Pagination } from '@/components/UI';
+import ItemsListPaginatedWrapper from '@/components/Customization/ItemsListPaginatedWrapper/ItemsListPaginatedWrapper';
+import { Button } from '@/components/UI';
 import { QUERY_KEYS } from '@/config';
 import { PaginationDto } from '@/dto';
-import { usePagination } from '@/hooks';
-import { useQuery } from '@tanstack/react-query';
+import { IItem } from '@/types';
 import { useSearchParams } from 'next/navigation';
 import { FieldValues, Path, PathValue, useFormContext } from 'react-hook-form';
-import styles from './ProductItemsList.module.scss';
 
 const ProductItemsList = <T extends FieldValues>() => {
 	const searchParams = useSearchParams();
@@ -20,14 +19,7 @@ const ProductItemsList = <T extends FieldValues>() => {
 		limit: 20,
 	};
 
-	const { data } = useQuery({
-		queryFn: () => getAllAvailableItems(params),
-		queryKey: QUERY_KEYS.customization.product.getAllAvailable(params),
-	});
-
 	const { setValue, watch } = useFormContext<T>();
-
-	const { currentPage, setPage } = usePagination(data?.meta.totalPages);
 
 	const itemId = watch('itemId' as Path<T>);
 
@@ -37,34 +29,26 @@ const ProductItemsList = <T extends FieldValues>() => {
 	};
 
 	return (
-		<div>
-			{data && data.items.length > 0 ? (
-				<div className={styles['product-items-list']}>
-					{data.items.map((item) => (
-						<ItemCard
-							key={item.id}
-							item={item}
-							actions={
-								<Button
-									variant={itemId === item.id ? 'contained' : 'secondary'}
-									fullWidth
-									onClick={() => handleSelect(item.id)}
-								>
-									{itemId === item.id ? 'Selected' : 'Select'}
-								</Button>
-							}
-						/>
-					))}
-					<Pagination
-						currentPage={currentPage}
-						onPageChange={setPage}
-						totalPages={data.meta.totalPages}
-					/>
-				</div>
-			) : (
-				<div>No Available Items</div>
+		<ItemsListPaginatedWrapper<IItem>
+			queryFn={() => getAllAvailableItems(params)}
+			queryKey={() => [
+				...QUERY_KEYS.customization.item.getAllAvailable(params),
+			]}
+			itemCard={(item) => (
+				<ItemCard
+					item={item}
+					actions={
+						<Button
+							variant={itemId === item.id ? 'contained' : 'secondary'}
+							fullWidth
+							onClick={() => handleSelect(item.id)}
+						>
+							{itemId === item.id ? 'Selected' : 'Select'}
+						</Button>
+					}
+				/>
 			)}
-		</div>
+		/>
 	);
 };