+++ client/src/components/Auth/EmailConfirmation/EmailConfirmation/EmailConfirmation.tsx
@@ -7,6 +7,9 @@ import { useMutation } from '@tanstack/react-query';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect } from 'react';
 import { toast } from 'react-toastify';
+import EmailConfirmationDefault from '../EmailConfirmationDefault/EmailConfirmationDefault';
+import EmailConfirmationProccesing from './EmailConfirmationStates/EmailConfirmationProccesing';
+import EmailConfirmationSuccess from './EmailConfirmationStates/EmailConfirmationSuccess';
 
 const EmailConfirmation = () => {
 	const searchParams = useSearchParams();
@@ -14,7 +17,7 @@ const EmailConfirmation = () => {
 	const router = useRouter();
 	const { user } = useAuth();
 
-	const { mutate, isSuccess } = useMutation({
+	const { mutate, isSuccess, isPending } = useMutation({
 		mutationFn: ({ token }: { token: string }) =>
 			validateVerificationToken(token),
 		mutationKey: QUERY_KEYS.auth.validateVerificationToken(
@@ -31,6 +34,7 @@ const EmailConfirmation = () => {
 				toast.info('Email already confirmed');
 			} else {
 				toast.error(error.message);
+				router.push(PAGES.EMAIL_CONFIRMATION);
 			}
 		},
 	});
@@ -45,7 +49,14 @@ const EmailConfirmation = () => {
 		}
 	}, [user, router]);
 
-	return <div>{isSuccess && 'Email confirmed'}</div>;
+	if (!user && !token) return <EmailConfirmationDefault />;
+
+	return (
+		<div>
+			{isPending && <EmailConfirmationProccesing />}
+			{isSuccess && <EmailConfirmationSuccess />}
+		</div>
+	);
 };
 
 export default EmailConfirmation;