+++ client/src/components/Layout/Header/Header.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import styles from './Header.module.scss';
-import HeaderEmailConfirmation from './HeaderEmailConfirmation/HeaderEmailConfirmation';
 import Nav from './Nav/Nav';
 
 const Header = () => {
@@ -10,7 +9,6 @@ const Header = () => {
 			<div className={`container ${styles['header__container']}`}>
 				<Nav />
 			</div>
-			<HeaderEmailConfirmation />
 		</header>
 	);
 };