+++ server/src/api/item/types/index.ts
@@ -1,0 +1,2 @@
+export * from './create-item-files.types';
+export * from './update-item-files.types';